package versions updated. sdkversion added. code refactoring
