+++ examples/light-bulb/light-bulb.js
@@ -1,26 +1,25 @@
-const { SinricPro, SinricProActions, raiseEvent, eventNames, SinricProUdp } = require('../../index'); // Use require('sinricpro'); if you are using NPM
+const { SinricPro, startSinricPro, raiseEvent, eventNames } = require('sinricpro'); 
+const APPKEY    = '';
+const APPSECRET = '';
+const light     = '';
+const deviceIds = [light]
 
-const appKey = ''; // d89f1***-****-****-****-************
-const secretKey = ''; // f44d1d31-1c19-****-****-9bc96c34b5bb-d19f42dd-****-****-****-************
-const device1 = ''; // 5d7e7d96069e275ea9******
-const deviceId = [device1]
-
-setPowerState = (deviceid, data) => {
+const setPowerState = (deviceid, data) => {
   console.log("Power state: ", deviceid, data);
   return true;
 }
 
-setColorTemperature = (deviceid, data) => {
+const setColorTemperature = (deviceid, data) => {
   console.log("Color temperature: ", deviceid, data);
   return true;
 }
 
-setColor = (deviceid, data) => {
+const setColor = (deviceid, data) => {
   console.log("Color: ", deviceid, data);
   return true;
 }
 
-setBrightness = (deviceid, data) => {
+const setBrightness = (deviceid, data) => {
   console.log("Brightness: ", deviceid, data);
   return true;
 }
@@ -32,6 +31,5 @@ const callbacks = {
   setColor
 };
 
-const sinricpro = new SinricPro(appKey, deviceId, secretKey, true);
-SinricProActions(sinricpro, callbacks);
-
+const sinricpro = new SinricPro(APPKEY, deviceIds, APPSECRET, true);
+startSinricPro(sinricpro, callbacks);
