+++ examples/switch/switch_observable.js
@@ -1,31 +1,11 @@
-const {
-  SinricPro, SinricProActions, raiseEvent, eventNames, SinricProUdp,
-} = require('../../index');
-const { SinricProActionsObservable } = require('../../lib/sinrcpro');
+const { SinricPro, startSinricPro, raiseEvent, eventNames } = require('sinricpro');
+const { startSinricProObservable } = require('sinricpro');
 
-/**
- *  Change to below code if your are using via npm
- *
- * const {SinricPro, SinricProActions, raiseEvent, eventNames,} = require('../index');
- *                                       ||
- *                                       ||
- *                                       ||
- *                                       ||
- *                                       ||
- *                                      \  /
- *                                       \/
- * const { SinricPro, SinricProActions, raiseEvent, eventNames,} = require('sinricpro');
- *
- *
- *
- */
-
-const appKey = ''; // d89f1***-****-****-****-************
-const secretKey = ''; // f44d1d31-1c19-****-****-9bc96c34b5bb-d19f42dd-****-****-****-************
-const device1 = ''; // 5d7e7d96069e275ea9******
-const device2 = ''; // 5d80ac5713fa175e99******
-const deviceId = [device1, device2];
-
+const appKey    = '';
+const secretKey = '';
+const device1   = '';
+const device2   = '';
+const deviceId  = [device1, device2];
 
 function setPowerState(deviceid, data) {
   console.log(deviceid, data);
@@ -42,7 +22,7 @@ const sinricpro = new SinricPro(appKey, deviceId, secretKey, true);
 let interval = null;
 
 // emit on connection is synched
-SinricProActionsObservable(sinricpro, callbacks).subscribe((emit) => {
+startSinricProObservable(sinricpro, callbacks).subscribe((emit) => {
   console.log(emit);
   const udp = new SinricProUdp(deviceId, secretKey);
