+++ lib/sinrcpro.js
@@ -11,17 +11,16 @@ const rx = require('rxjs');
 const callbackHandler = require('./cbhandler');
 const { raiseEvent, eventNames } = require('./events/event_handler');
 const SinricProUdp = require('./sinricproudp');
-
-const VERSION = '2.2.2';
-
+const sdkVersion = require('./../package.json').version;
 class SinricPro extends Websocket {
+
   constructor(appkey, deviceids, secretKey, restoreStates) {
     super('ws://ws.sinric.pro', {
       headers: {
         appkey,
         deviceids: deviceids.join(';'),
         platform: 'nodejs',
-        version: VERSION,
+        sdkversion: sdkVersion,
         restoredevicestates: restoreStates,
       },
     });
@@ -37,7 +36,7 @@ class SinricPro extends Websocket {
   }
 }
 
-const SinricProActions = (client, callbacks) => {
+const startSinricPro = (client, callbacks) => {
   client.on('open', () => {
     console.log('Connected to Sinric Pro');
   });
@@ -66,7 +65,7 @@ const SinricProActions = (client, callbacks) => {
   });
 };
 
-const SinricProActionsObservable = (client, callbacks) => new rx.Observable((observer) => {
+const startSinricProObservable = (client, callbacks) => new rx.Observable((observer) => {
   client.on('open', () => {
     console.log('Connected');
   });
@@ -103,8 +102,8 @@ const SinricProActionsObservable = (client, callbacks) => new rx.Observable((obs
 
 module.exports = {
   SinricPro,
-  SinricProActions,
-  SinricProActionsObservable,
+  startSinricPro,
+  startSinricProObservable,
   raiseEvent,
   eventNames,
   SinricProUdp,
